CAN-NER/requirements.txt
numpy==1.14.0
pytorch==1.3.1